Refactor GenSym.
